small improvements and refactor
